Falppy Bird In Python
In this project, we design and implement a Flappy Bird-like video game by using Python Programming Language. In this game, the player can control the vertical movement of the bird (every pressing on the keyboard makes the bird leap upward for a little bit, and the bird will fall freely without control.
Flappy Bird is a game where you tap the enter button to make the bird fly. If the bird hits the obstacles or the edges of the screen, the game ends and the player needs to restart. we have made a computer version of the game where the bird will be controlled using the Enter key and mouse buttons(Also TouchPad). We will be using Python language for writing the code. We will also be using Pygame which is a cross-platform set of Python modules designed for writing video games.